iwlwifi: make IWLWIFI_PLATFORM_DATA automatic
The symbol IWLWIFI_PLATFORM_DATA currently has to be carefully
selected manually - if it's selected erroneously then the build
will fail, and if it's not selected when it should be the power
GPIOs won't be used.

Change it to be a non-selectable symbol that always defaults to
yes if the driver is built on a platform that has an appropriate
platform power management implementation.

type=cleanup

Change-Id: I3c63b068f2dbed1ac7863b91cc9f0268ebabb01c
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/59450
Reviewed-by: Aymen Zayet <aymen.zayet@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 66459c893ee72d6eb1cee929ddf77a434097864c