iwlwifi: pcie: rx: add emergency allocation path
On certain environments, such as virtualized, scheduling
issues with the background allocator might occur, causing
the workqueue to not be scheduled often enough.
This can leads to allocation requests not satisfied, and
eventually the queue will be empty and crash.

Add an emergency path in the interrupt handler.
Emergency will start when allocator owns half of the queue.
While in emergency the queue will stop providing RBDs to the
allocator, continue claiming previous requests and allocate
in the meanwhile each 8 emptied buffers in the foreground.
Emergency will be turned off once allocator's owned RBDs
are less than third of the queue.
This path is very unlikely to be used.

type=feature

Change-Id: I14c1656f0e7058a8fc62d95256e4c4a1a66b7c4f
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/59747
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: b9f7b812b2e05f101798e18b0bc03a878ad72121