iwlwifi: pcie: don't warn on long MPDUs when supported
In iwlmvm firmwares, the Byte count written in the scheduler
byte count table is in DWORDs and not in bytes.
We should check that this value fits in the 12 bits and
the value can be either in bits of in DWORD or bytes
depending on the firmware. Check the value after the
translation to DWORDs is done (if needed).

type=feature

Change-Id: I8de0ee68087b0404f96e340d33640c24df900b71
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
This is really needed for A-MSDU only (and only for >4KB)
Reviewed-on: https://gerrit.rds.intel.com/r/59488
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 9c0cb33ae6b584f1239fe16ab839c3c7467b946c