[NOUPSTREAM] mac80211: Call a different function when tx latency crossed a thrshld
Until now the old API was called (drv_retrieve_monitor_logs) to notify the
driver that the latency of a Tx packet crossed a configured threshold.

Due to the addition of a new API(drv_event_callback) that serves many other
flows, change the API that is called to be the new one.

Change the function that is called in case that a Tx packet crossed a
thershold to the newer API.

type=feature

Change-Id: I217d44891afd535d33e41092fd285760ef63ce7f
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/59359
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 867b25b9a062c78a7748f3090ccc99f923ae16c2