iwlwifi: pcie: provide a way to stop configuration if it is forbidden
The firmware debug infrastructure allows the user to
provide a firmware that will toggle a few registers to
configure the debugging capabilities.
On certain devices, certain operations are forbidden.
Executing a forbidden operation will cause the hardware to
die in a way that only driver unload / load will bring it
back to life.
Fortunately, there is a way to know in advance if those
operations will be accepted by the device. This is where
the new PRPH_BLOCKBIT operation plays its role. If the bit
X from PRPH register Y is set, then we should prevent any
further register configuration. When that happens, drop a
line in the kernel log since this is really an error state:
the user won't have his device configured as he expected.

Other debugging configurations (such as destination
configuration for the monitor) will take place in any case.

type=feature

Change-Id: Iab1ec23344dd12e22889834d476bc440cc933d2e
Signed-off-by: Haim Dreyfuss <haim.dreyfuss@intel.com>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
The primary purpose of that is to add a possibility to bypass secure
boot. WRT will add the register operations needed to bypass secure
boot, but that can fail.
Check that we can technically override secure boot with this new
PRPH_BLOCKBIT operation. If the bit is set, stop running the register
ops that would have overridden the secure boot.
Reviewed-on: https://gerrit.rds.intel.com/r/60348
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: da199b2e75e98bdd6a5c87775b1ac349cb77ca12