[NOUPSTREAM] iwlwifi: sdio: provide a way to stop configuration if it is forbidden
The firmware debug infrastructure allows the user to
provide a firmware that will toggle a few registers to
configure the debugging capabilities.
On certain devices, certain operations are forbidden.
Executing a forbidden operation will cause the hardware to
die in a way that only driver unload / load will bring it
back to life.
Fortunately, there is a way to know in advance if those
operations will be accepted by the device. This is where
the new PRPH_BLOCKBIT operation plays its role. If the bit
X from PRPH register Y is set, then we should prevent any
further register configuration. When that happens, drop a
line in the kernel log since this is really an error state:
the user won't have his device configured as he expected.

The primary purpose of that is to add a possibility to bypass secure
boot. WRT will add the register operations needed to bypass secure
boot, but that can fail.
Check that we can technically override secure boot with this new
PRPH_BLOCKBIT operation. If the bit is set, stop running the register
ops that would have overridden the secure boot.

type=feature

Change-Id: I72f9ee98bd1136e1e81cdd7edf410f39812fe7ec
Signed-off-by: Haim Dreyfuss <haim.dreyfuss@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/60580
Tested-by: IWL Jenkins
Reviewed-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 3a73655b34f83b64b9e874b2539f90e735528046