[BUGFIX] mac80211: protect non-HT BSS when HT TDLS traffic exists
HT TDLS traffic should be protected in a non-HT BSS to avoid
collisions. Therefore, when TDLS peers join/leave, check if
protection is (now) needed and set the ht_operation_mode of
the virtual interface according to the HT capabilities of the
TDLS peer(s).

This works because a non-HT BSS connection never sets (or
otherwise uses) the ht_operation_mode; it just means that
drivers must be aware that this field applies to all HT
traffic for this virtual interface, not just the traffic
within the BSS. Document that.

type=bugfix
bug=cq:MWG100239181
fixes=unknown

Change-Id: Iaf869ad91ee4fbaa782c454a3e488e4493e4c9fb
Signed-off-by: Avri Altman <avri.altman@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/60749
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: 40692f3a8f6285f165c04ebb86c89aa847340035