[BUGFIX] iwlwifi: correctly size command string arrays
The command string array is assumed to be indexable by a full u8, so it
must have 256 entries (0-255), not just 255. A recent firmware change
(apparently) started using the command 0xff for a notification to the
host, causing the driver to crash in debug message/tracing code.

For safety, fix this for all the opmodes, even if only the recently
changed firmware is able to trigger this.

type=bugfix
bug=cq:MWG100244717
fixes=unknown

Change-Id: If5a13bba44a4492a832a7e3c40fe78fa26738d6a
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/1906
Tested-by: Griva, Gil <gil.griva@intel.com>
x-iwlwifi-stack-dev: 48cb03eff830147dac9035c4fb61077698e07f1b