[BUGFIX] iwlwifi: trans_slave: take wake_lock on runtime_resume failure
On runtime_resume we exit d0i3 and take back the wakelock
(which will be released once we'll runtime_suspend again).

However, if we fail to exit d0i3 (e.g. due to dead bus),
we exit prematurely and don't take the wakelock.

This ends up in a situation where the runtime_pm is no
longer functional until the upcoming recovery (since we returned
error from the runtime_resume callback), but the wakelock
is not taken.

This in turn allows the OS the suspend the host, although
the driver is not ready for it.

Solve it by taking the wakelock regardless the d0i3 exit
status. In case of error, a further recovery should release
the wakelock after the driver enters d0i3 again.

type=bugfix
bug=not-tracked
fixes=unknown

Change-Id: I58b971e69e1b4be81c681a27bc02be4bb9aedd14
Signed-off-by: Eliad Peller <eliadx.peller@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/3082
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
Reviewed-by: Greenman, Gregory <gregory.greenman@intel.com>
x-iwlwifi-stack-dev: 548fc6b52bbc2c1baf43794b248d87b1c60234af