[BUGFIX] iwlwifi: nvm: fix up phy section when reading it
This is a workaround to an OTP bug. In Series 8000 1x1, the OTP
0xA052 defines 2x2 antenna configuration. This workaround overrides
the decision based on HW id and on MIMO disabled bit which is
correct in the OTP and set to disabled. This fixes the previous
workaround "force 1x1 antenna in Series 8000".

type=bugfix
bug=not-tracked
fixes=unknown

Change-Id: Iccc0bdbe6982e92e2680150e876aebaed2f64837
Signed-off-by: Moshe Harel <moshe.harel@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/3677
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 55ed2f07994583181813fbea576bd9b3370dc08d