[NOUPSTREAM] mac80211: fix variable shadowing warning in tx-latency
Sparse complains that the variable 'size' is being shadowed, this is
due to upstream introducing it in an if branch while our code had it
in the function itself for the tx-latency code. Move the latter one
into another if branch to avoid the warning.

type=cleanup

Change-Id: Ibd01b9d66de60d5d59a68245639262511e380816
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/4219
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 6c5a93a2c40ea6eefd0c77a451aa110ecea99c78