[BUGFIX] iwlwifi: mvm: fix message allocation for significant change event
The firmware supports reporting 8 AP's and a history of 8 RSSI
samples for each AP. The report for each AP consists of the BSSID,
channel, and RSSI history, which together with nesting and headers
takes ~100 bytes. So the 200 bytes allocated for this event are not
enough, which may result in not sending the event when it includes
more than 2 AP's.
Fix this by allocating 1024 bytes for this event.

type=bugfix
bug=not-tracked
fixes=Ia2aab62ccbbae8d7785532fb75ee87a02377c297

Change-Id: Id9658804cc12d9d9594cafe63627cba26120a5d7
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/6056
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 77dad628144dafa03f438c81678b92daf2cdebcc