[BUGFIX] iwlwifi: sdio: change when the SCD_GP_CTRL is set
There was a bug where the driver would update the SCD_GP_CTRL
periphery after the firmware had already loaded. This caused
the update of the periphery not to be recognized by the firmware.
Thus, in this case, instead of 31 queues, the firmware thought
it had only 20 queues.
Moved the update of the periphery to be before the firmware
is loaded in iwl_sdio_tx_init instead of iwl_sdio_tx_start
(just like it is in pcie).

type=bugfix
fixes=I27218a9f0e348f0c4376a4ba0f17a5be7ae42304
bug=cq:MWG100249765

Change-Id: Ia2d0220eece28239330e608d1f7833d98e2e4518
Signed-off-by: Oren Givon <oren.givon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/8496
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 1ce884824679552906663652dec171c50704cabe