iwlwifi: sdio: remove unnecessary code
As the SDTM read optimisation (omitting the write) is currently
completely disabled, there's no need to set the hw_rev to 0 to
avoid the optimisation.

If this needs to be added back it should also be done explicitly
rather than on a hw_rev check - "is_irq_write_necessary" or so.

type=cleanup

Change-Id: Ia9e58eef624f306bf34ae28b0db62c523a14f354
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/8886
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Kaufman, Liad <liad.kaufman@intel.com>
x-iwlwifi-stack-dev: d57f7890495ccf613626495a18071fcb2dfe3ae4