nl80211: add an option to set driver action for beacon loss
When beacon loss is detected the driver will try to probe the AP and
if the AP does not respond, it will disconnect. However, userspace
may not want to disconnect immediately but look for a roaming
candidate first.
So add a flag to connect and association commands that indicates that
for the current connection the driver should not disconnect when
beacon loss is detected but only send an event to userspace.

type=feature

Change-Id: If8374e3d58057aabe1f9c30e3d838586fa00ee9b
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/9662
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 9a2427c9a299da02dbc582afb0c83db051ce1387