[NOUPSTREAM] iwlwifi: pcie: add pm_prepare and pm_complete ops
With these ops, we can know when we are about to enter system suspend.
This allows us to set a flag indicating that we are suspending and use
it to prevent the runtime resume operations (which is called as part
of the suspend flow) from grabbing the wakelock and aborting the
system suspend entry.

type=feature

Change-Id: I70c0686e071a52579e178051eab553a805c9f4dd
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/15839
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: a789886c8c15cc777ca6d98de4046be61788eca2