[BUGFIX] mac80211: Properly handle HW ROC expired
In case of HW ROC, when the driver reports that the ROC expired,
it is not sufficient to purge the ROCs based on the remaining
time, as it possible that the device finished the ROC session
before the actual requested duration.

To handle such cases, in case of ROC expired notification from
the driver, complete all the ROCs which are marked with hw_begun,
regardless of the remaining duration.

type=bugfix
bug=not-tracked
fixes=Ic787c32108cf424513f0d1afdbe5edff513c6daf

Change-Id: I86d7a2d1a6560dc67b5fa1bd6fadb6d8b1edbe2d
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/11406
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 15fb85a3e89aa037c73b3bfc636dde6a8d4c29b2