[BUGFIX] iwlwifi: pcie: properly configure the debug buffer size for 8000
8000 device family has a new debug engine that needs to be
configured differently than 7000's.
The debug engine's DMA works in chunks of memory and the
size of the buffer really means the start of the last
chunk. Since one chunk is 256-byte long, we should
configure the device to write to buffer_size - 256.
This fixes a situation were the device would write to
memory it is not allowed to access.

type=bugfix
bug=not-tracked
fixes=unknown

Change-Id: Iff0c91c4ec20c5597097c115c960328d4caeba97
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/14977
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Kaufman, Liad <liad.kaufman@intel.com>
x-iwlwifi-stack-dev: ef08f9abcdd18bd5c215bdd76e4af29cdb62e46a