iwlwifi: mvm: keep track of tid associated with each queue
When sending the SCD_QUEUE_CONFIG command, the queue is
associated to a specific TID. If later there is a need to
use this TID on a different queue instead, it first needs to
be unassociated from the first queue.

Keep track for every queue what TID is associated with it.

type=feature

Change-Id: I9d69ee8df840e4ffea6b17e96ca1fa33535985e9
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/40683
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Coelho, Luciano <luciano.coelho@intel.com>
x-iwlwifi-stack-dev: 791ff808df62c86e46b0d1029786b9628293cbb6