cfg80211: add FTM capabilties advertisement mechanism
Publish a set of FTM capabilities supported by the wiphy.
All requests that will ask for a not supported capability
will be refused.

Note: though this patch is cfg82011, it includes a small portion of
mvm in order not to break compilation

type=feature

Change-Id: Idce38b40665f75aca00d79a5225057da6ea4da98
Signed-off-by: Beni Lev <beni.lev@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/20275
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 377a30045dbecb49f5884572ed66ce9481ff85c9