iwlwifi: store fw memory segments length and addresses in run-time
Currently reading the fw memory segments is done according to
addresses and data length that are hard-coded.
Lately a new tlv was appended to the ucode, that contains
the data type, length and address.
Parse this tlv, and in run-time store the memory segments length
and addresses that would be dumped upon a fw error.

type=feature

Change-Id: I9bad784ac52a5fe430957dc70cd0393d74b2da70
Signed-off-by: Golan Ben-Ami <golan.ben.ami@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/23758
Reviewed-by: Grumbach, Emmanuel <emmanuel.grumbach@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: f030026c3103ce0551cd6306221ec6fb1303171e